refactor: rename better-auth-config to better-auth and update imports
