--- auth.ts
@@ -1,2 +1,0 @@
-import { auth } from "~libs/auth/better-auth-config";
-export default auth;
