+++ package.json
@@ -11,7 +11,8 @@
     "check": "astro check",
     "db:generate": "drizzle-kit generate",
     "db:push": "drizzle-kit push",
-    "db:studio": "drizzle-kit studio"
+    "db:studio": "drizzle-kit studio",
+    "auth:generate": "npx @better-auth/cli generate --config src/libs/auth/better-auth.ts --output db/schemas/auth-schema.ts --y"
   },
   "dependencies": {
     "@astrojs/check": "^0.9.4",
